requirements.txt
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
fastapi==0.110.0
huey==2.5.0
redis==5.0.2
uvicorn==0.27.1
